requirements.txt
streamlit==1.56.0
duckdb==1.5.1
langchain==1.2.15
langchain-community==0.4.1
langchain-core==1.3.0
langchain-mistralai==1.1.2
langfuse==4.3.1
matplotlib==3.10.8
pandas==3.0.1
python-dotenv==1.2.2
